refacto(scopes): split variables module into multiple files
