+++ packages/scopes/src/Resource.ts
@@ -1,11 +1,12 @@
-import type { Var } from './Var'
-
-export type Resource<T> = {
+export type Resource<T = any> = {
   value: T
-  unmount?: Var.Unmount
+  unmount?: Resource.Unmount
+}
+export namespace Resource {
+  export type Unmount = () => void | Promise<void>
 }
 
-export const of = <T>(value: T, unmount?: Var.Unmount): Resource<T> => ({
+export const of = <T>(value: T, unmount?: Resource.Unmount): Resource<T> => ({
   value,
   unmount
 })
