+++ packages/scopes/src/index.ts
@@ -1,4 +1,4 @@
 export { Context } from './types'
 export { Scope } from './Scope'
-export { Var } from './Var'
+export { Var } from './variables'
 export { Resource } from './Resource'
